test: keep retry regressions handwritten
